Passing densite_pouss to simgle precision

+++ src/modules.f90
@@ -437,7 +437,7 @@ module opacity
   real, dimension(:,:,:), allocatable :: kabs_nLTE_CDF, kabs_nRE_CDF ! 0:n_grains, n_cells, n_lambda
   real(kind=dp), dimension(:,:), allocatable :: emissivite_dust ! emissivite en SI (pour mol)
 
-  real(kind=dp), dimension(:,:), allocatable :: densite_pouss ! n_grains, n_cells en part.cm-3
+  real, dimension(:,:), allocatable :: densite_pouss ! n_grains, n_cells en part.cm-3
   integer :: icell_not_empty
 
   real, dimension(:,:,:), allocatable :: ksca_CDF ! 0:n_grains, n_cells, n_lambda
